Enable to set environment variables for db in test
